ExtensionDescriptorMerger
