+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/mergers/v2/ExtensionDescriptorMerger.java
@@ -1,8 +1,12 @@
 package com.sap.cloud.lm.sl.mta.mergers.v2;
 
+import java.util.Map;
+
 import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler;
 import com.sap.cloud.lm.sl.mta.model.ElementContext;
-import com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor;
+import com.sap.cloud.lm.sl.mta.model.PropertiesContainer;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
+import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionRequiredDependency;
@@ -12,16 +16,20 @@ import com.sap.cloud.lm.sl.mta.model.v2.RequiredDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.Resource;
 import com.sap.cloud.lm.sl.mta.util.PropertiesUtil;
 
-public class ExtensionDescriptorMerger extends com.sap.cloud.lm.sl.mta.mergers.v1.ExtensionDescriptorMerger {
+public class ExtensionDescriptorMerger extends Visitor {
+
+    protected final ExtensionDescriptor extensionDescriptor;
+    protected final DescriptorHandler handler;
 
     public ExtensionDescriptorMerger(ExtensionDescriptor extensionDescriptor, DescriptorHandler handler) {
-        super(extensionDescriptor, handler);
+        this.extensionDescriptor = extensionDescriptor;
+        this.handler = handler;
     }
 
     @Override
     public void visit(ElementContext context, DeploymentDescriptor descriptor) {
-        com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor descriptorV2 = (com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor) descriptor;
-        ExtensionDescriptor extensionDescriptorV2 = (ExtensionDescriptor) extensionDescriptor;
+        DeploymentDescriptor descriptorV2 = descriptor;
+        ExtensionDescriptor extensionDescriptorV2 = extensionDescriptor;
         descriptorV2
             .setParameters(PropertiesUtil.mergeExtensionProperties(descriptorV2.getParameters(), extensionDescriptorV2.getParameters()));
     }
@@ -44,24 +52,18 @@ public class ExtensionDescriptorMerger extends com.sap.cloud.lm.sl.mta.mergers.v
             .setProperties(PropertiesUtil.mergeExtensionProperties(requiredDependency.getProperties(), extension.getProperties()));
     }
 
-    @Override
-    public void merge(com.sap.cloud.lm.sl.mta.model.v1.Module module, com.sap.cloud.lm.sl.mta.model.v1.ExtensionModule extension) {
-        merge((Module) module, (ExtensionModule) extension);
-    }
-
-    private void merge(Module module, ExtensionModule extension) {
+    public void merge(Module module, ExtensionModule extension) {
+        module.setProperties(mergeProperties(module, extension));
         module.setParameters(PropertiesUtil.mergeExtensionProperties(module.getParameters(), extension.getParameters()));
-        super.merge(module, extension);
     }
 
-    @Override
-    public void merge(com.sap.cloud.lm.sl.mta.model.v1.Resource resource,
-        com.sap.cloud.lm.sl.mta.model.v1.ExtensionResource extension) {
-        merge((Resource) resource, (ExtensionResource) extension);
+    protected Map<String, Object> mergeProperties(PropertiesContainer container, PropertiesContainer extensionContainer) {
+        return PropertiesUtil.mergeExtensionProperties(container.getProperties(), extensionContainer.getProperties());
     }
 
-    private void merge(Resource resource, ExtensionResource extension) {
+    public void merge(Resource resource, ExtensionResource extension) {
+        resource.setProperties(mergeProperties(resource, extension));
         resource.setParameters(PropertiesUtil.mergeExtensionProperties(resource.getParameters(), extension.getParameters()));
-        super.merge(resource, extension);
     }
+
 }
